feat: separate code in helper into several files
