+++ app.js
@@ -1,4 +1,6 @@
 'use strict';
+
+const path = require('path');
 // app.js
 module.exports = app => {
   /**
@@ -9,9 +11,11 @@ module.exports = app => {
     console.log(app.tips);
 
     // 删除 socket 所有连接;处理未完成的番茄钟( TODO )
-
   });
 
   // true: 可以直接在页面引用 /socket.io/socket.io.js
   app.io.serveClient(false);
+
+  const directory = path.join(app.config.baseDir, 'app/util');
+  app.loader.loadToApp(directory, 'util');
 };
