+++ app/controller/user.js
@@ -22,7 +22,7 @@ class UserController extends BaseController {
         const password = users[0].password;
         const verifyPWD = await users[0].comparePassword(ctx.request.body.password);
         if (password === ctx.request.body.password || verifyPWD) {
-            const token = ctx.helper.tokenService.createToken(user);
+            const token = app.util.jwt.tokenService.createToken(user);
             ctx.logger.info(user, token);
             await app.redis.set(token, JSON.stringify(user), 'EX', 3 * 24 * 60 * 60);
             ctx.body = {
